backend/requirements.txt
flask==3.0.3
flask_sqlalchemy==3.1.1
flask-jwt-extended==4.6.0
python-dotenv==1.0.1
